refactor: remove the chain_selection module
