--- crates/amaru-consensus/src/consensus/chain_selection.rs
@@ -1,66 +1,0 @@
-// Copyright 2024 PRAGMA
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use amaru_kernel::{peer::Peer, Point};
-use amaru_ouroboros_traits::is_header::IsHeader;
-use pallas_crypto::hash::Hash;
-use std::fmt::Debug;
-
-pub const DEFAULT_MAXIMUM_FRAGMENT_LENGTH: usize = 2160;
-
-/// Definition of a fork.
-///
-/// FIXME: The peer should not be needed here, as the fork should be
-/// comprised of known blocks. It is only needed to download the blocks
-/// we don't currently store.
-#[derive(Debug, PartialEq)]
-pub struct Fork<H: IsHeader> {
-    pub peer: Peer,
-    pub rollback_point: Point,
-    pub fork: Vec<H>,
-}
-
-/// The outcome of the chain selection process in  case of
-/// roll forward.
-#[derive(Debug, PartialEq)]
-pub enum ForwardChainSelection<H: IsHeader> {
-    /// The current best chain has been extended with a (single) new header.
-    NewTip { peer: Peer, tip: H },
-
-    /// The current best chain is unchanged.
-    NoChange,
-
-    /// The current best chain has switched to given fork.
-    SwitchToFork(Fork<H>),
-}
-
-/// The outcome of the chain selection process in case of rollback
-#[derive(Debug, PartialEq)]
-pub enum RollbackChainSelection<H: IsHeader> {
-    /// The current best chain has been rolled back to the given hash.
-    RollbackTo(Hash<32>),
-
-    /// The current best chain has switched to given fork.
-    SwitchToFork(Fork<H>),
-
-    /// The peer tried to rollback beyond the limit
-    RollbackBeyondLimit {
-        peer: Peer,
-        rollback_point: Hash<32>,
-        max_point: Hash<32>,
-    },
-
-    /// The current best chain as not changed
-    NoChange,
-}
